use the same component
