+++ packages/mui-material/src/Button/Button.js
@@ -6,9 +6,9 @@ import resolveProps from '@mui/utils/resolveProps';
 import composeClasses from '@mui/utils/composeClasses';
 import { alpha } from '@mui/system/colorManipulator';
 import rootShouldForwardProp from '../styles/rootShouldForwardProp';
-import { styled } from '../zero-styled';
+import { styled, useTheme } from '../zero-styled';
 import { useDefaultProps } from '../DefaultPropsProvider';
-import ButtonBase from '../ButtonBase';
+import ButtonBase, {buttonBaseClasses} from '../ButtonBase';
 import capitalize from '../utils/capitalize';
 import buttonClasses, { getButtonUtilityClass } from './buttonClasses';
 import ButtonGroupContext from '../ButtonGroup/ButtonGroupContext';
@@ -345,12 +345,326 @@ const ButtonEndIcon = styled('span', {
   ],
 }));
 
+export const ButtonRootMD3 = styled(ButtonBase, {
+  name: 'MuiButton',
+  slot: 'Root',
+  shouldForwardProp: (prop) => rootShouldForwardProp(prop) || prop === 'classes',
+  overridesResolver: (props, styles) => {
+    const { ownerState } = props;
+
+    return [
+      styles.root,
+      styles[ownerState.variant],
+      styles[`${ownerState.variant}${capitalize(ownerState.color)}`],
+      styles[`size${capitalize(ownerState.size)}`],
+      styles[`${ownerState.variant}Size${capitalize(ownerState.size)}`],
+      ownerState.color === 'inherit' && styles.colorInherit,
+      ownerState.disableElevation && styles.disableElevation,
+      ownerState.fullWidth && styles.fullWidth,
+    ];
+  },
+})(({ ownerState, theme }) => {
+  const tokens = theme.vars || theme;
+
+  const containerColor = { 
+    elevated: `linear-gradient(0deg, rgba(103, 80, 164, 0.05), rgba(103, 80, 164, 0.05)), ${tokens.sys.color.surface}`,
+    filled: tokens.sys.color[ownerState.color ?? 'primary'],
+    filledTonal: tokens.sys.color.secondaryContainer,
+    outlined: 'transparent',
+    text: 'transparent',
+  };
+
+  const labelTextColor = {
+    elevated: tokens.sys.color.primary,
+    filled:
+      tokens.sys.color[
+        `on${capitalize(ownerState.color ?? 'primary')}`
+      ],
+    filledTonal: tokens.sys.color.onSecondaryContainer,
+    outlined: tokens.sys.color[ownerState.color ?? 'primary'],
+    text: tokens.sys.color[ownerState.color ?? 'primary'],
+  };
+
+  const disabledContainerColor = {
+    elevated: theme.vars
+      ? `rgba(${theme.vars.sys.color.onSurfaceChannel} / 0.12)`
+      : alpha(theme.sys.color.onSurface, 0.12),
+    filled: theme.vars
+      ? `rgba(${theme.vars.sys.color.onSurfaceChannel} / 0.12)`
+      : alpha(theme.sys.color.onSurface, 0.12),
+    filledTonal: theme.vars
+      ? `rgba(${theme.vars.sys.color.onSurfaceChannel} / 0.12)`
+      : alpha(theme.sys.color.onSurface, 0.12),
+    outlined: 'transparent',
+    text: 'transparent',
+  };
+
+  const hoveredContainerColor = {
+    elevated: theme.vars
+      ? `rgba(${tokens.sys.color.primaryChannel} / ${tokens.sys.state.hover.stateLayerOpacity})`
+      : alpha(theme.sys.color.primary, theme.sys.state.hover.stateLayerOpacity),
+    filled: theme.vars
+      ? `rgba(${tokens.sys.color[`${ownerState.color ?? 'primary'}Channel`]} / calc(1 - ${
+          tokens.sys.state.hover.stateLayerOpacity
+        }))`
+      : alpha(
+          theme.sys.color[ownerState.color ?? 'primary'],
+          1 - theme.sys.state.hover.stateLayerOpacity,
+        ),
+    filledTonal: theme.vars
+      ? `rgba(${tokens.sys.color.secondaryContainerChannel} / calc(1 - ${tokens.sys.state.hover.stateLayerOpacity}))`
+      : alpha(theme.sys.color.secondaryContainer, 1 - theme.sys.state.hover.stateLayerOpacity),
+    outlined: theme.vars
+      ? `rgba(${tokens.sys.color[`${ownerState.color ?? 'primary'}Channel`]} / ${
+          tokens.sys.state.hover.stateLayerOpacity
+        })`
+      : alpha(
+          theme.sys.color[ownerState.color ?? 'primary'],
+          theme.sys.state.hover.stateLayerOpacity,
+        ),
+    text: theme.vars
+      ? `rgba(${tokens.sys.color[`${ownerState.color ?? 'primary'}Channel`]} / ${
+          tokens.sys.state.hover.stateLayerOpacity
+        })`
+      : alpha(
+          theme.sys.color[ownerState.color ?? 'primary'],
+          theme.sys.state.hover.stateLayerOpacity,
+        ),
+  };
+
+  const pressedContainerColor = {
+    elevated: theme.vars
+      ? `rgba(${tokens.sys.color.primaryChannel} / ${tokens.sys.state.pressed.stateLayerOpacity})`
+      : alpha(theme.sys.color.primary, theme.sys.state.pressed.stateLayerOpacity),
+    filled: theme.vars
+      ? `rgba(${tokens.sys.color[`${ownerState.color ?? 'primary'}Channel`]} / calc(1 - ${
+          tokens.sys.state.pressed.stateLayerOpacity
+        }))`
+      : alpha(
+          theme.sys.color[ownerState.color ?? 'primary'],
+          1 - theme.sys.state.pressed.stateLayerOpacity,
+        ),
+    filledTonal: theme.vars
+      ? `rgba(${tokens.sys.color.secondaryContainerChannel} / calc(1 - ${tokens.sys.state.pressed.stateLayerOpacity}))`
+      : alpha(theme.sys.color.secondaryContainer, 1 - theme.sys.state.pressed.stateLayerOpacity),
+    outlined: theme.vars
+      ? `rgba(${tokens.sys.color[`${ownerState.color ?? 'primary'}Channel`]} / ${
+          tokens.sys.state.pressed.stateLayerOpacity
+        })`
+      : alpha(
+          theme.sys.color[ownerState.color ?? 'primary'],
+          theme.sys.state.pressed.stateLayerOpacity,
+        ),
+    text: theme.vars
+      ? `rgba(${tokens.sys.color[`${ownerState.color ?? 'primary'}Channel`]} / ${
+          tokens.sys.state.pressed.stateLayerOpacity
+        })`
+      : alpha(
+          theme.sys.color[ownerState.color ?? 'primary'],
+          theme.sys.state.pressed.stateLayerOpacity,
+        ),
+  };
+
+  const focusedContainerColor = {
+    elevated: theme.vars
+      ? `rgba(${tokens.sys.color.primaryChannel} / ${tokens.sys.state.focus.stateLayerOpacity})`
+      : alpha(theme.sys.color.primary, theme.sys.state.focus.stateLayerOpacity),
+    filled: theme.vars
+      ? `rgba(${tokens.sys.color[`${ownerState.color ?? 'primary'}Channel`]} / calc(1 - ${
+          tokens.sys.state.focus.stateLayerOpacity
+        }))`
+      : alpha(
+          theme.sys.color[ownerState.color ?? 'primary'],
+          1 - theme.sys.state.focus.stateLayerOpacity,
+        ),
+    // According to the spec, this should be: secondaryContainerChannel / 1 - focusStateLayerOpacity, but this doesn't have the enough contrast
+    filledTonal: theme.vars
+      ? `rgba(${tokens.sys.color.primaryChannel} / 0.3)`
+      : alpha(theme.sys.color.primary, 0.3),
+    outlined: theme.vars
+      ? `rgba(${tokens.sys.color[`${ownerState.color ?? 'primary'}Channel`]} / ${
+          tokens.sys.state.focus.stateLayerOpacity
+        })`
+      : alpha(
+          theme.sys.color[ownerState.color ?? 'primary'],
+          theme.sys.state.focus.stateLayerOpacity,
+        ),
+    text: theme.vars
+      ? `rgba(${tokens.sys.color[`${ownerState.color ?? 'primary'}Channel`]} / ${
+          tokens.sys.state.focus.stateLayerOpacity
+        })`
+      : alpha(
+          theme.sys.color[ownerState.color ?? 'primary'],
+          theme.sys.state.focus.stateLayerOpacity,
+        ),
+  };
+
+  const containerElevation = {
+    elevated: tokens.sys.elevation[1],
+    filled: tokens.sys.elevation[0],
+    filledTonal: tokens.sys.elevation[0],
+    outlined: tokens.sys.elevation[0],
+    text: tokens.sys.elevation[0],
+  };
+
+  const hoveredContainerElevation = {
+    elevated: tokens.sys.elevation[2],
+    filled: tokens.sys.elevation[1],
+    filledTonal: tokens.sys.elevation[1],
+    outlined: tokens.sys.elevation[0],
+    text: tokens.sys.elevation[0],
+  };
+
+  const focusedContainerElevation = {
+    elevated: tokens.sys.elevation[1],
+    filled: tokens.sys.elevation[0],
+    filledTonal: tokens.sys.elevation[0],
+    outlined: tokens.sys.elevation[0],
+    text: tokens.sys.elevation[0],
+  };
+
+  const pressedContainerElevation = {
+    elevated: tokens.sys.elevation[1],
+    filled: tokens.sys.elevation[0],
+    filledTonal: tokens.sys.elevation[0],
+    outlined: tokens.sys.elevation[0],
+    text: tokens.sys.elevation[0],
+  };
+
+  const disabledLabelTextColor = theme.vars
+    ? `rgba(${theme.vars.sys.color.onSurfaceChannel} / 0.38)`
+    : alpha(theme.sys.color.onSurface, 0.38);
+
+  const letterSpacing = `${
+    theme.sys.typescale.label.large.tracking / theme.sys.typescale.label.large.size
+  }rem`;
+
+  const borderRadiusValue = tokens.sys.shape.corner.full;
+  const borderRadius = Number.isNaN(Number(borderRadiusValue))
+    ? borderRadiusValue
+    : `${borderRadiusValue}px`;
+
+  return {
+    // Icon variables default values
+    '--md-comp-button-icon-color': labelTextColor[ownerState.variant ?? 'text'],
+    '--md-comp-button-hovered-icon-color': labelTextColor[ownerState.variant ?? 'text'], // same as default
+    '--md-comp-button-pressed-icon-color': labelTextColor[ownerState.variant ?? 'text'], // same as default
+    '--md-comp-button-focused-icon-color': labelTextColor[ownerState.variant ?? 'text'], // same as default
+    '--md-comp-button-disabled-icon-color': disabledLabelTextColor,
+    padding: '10px 24px',
+    minWidth: 64,
+    letterSpacing,
+    transition: theme.sys.motion.create(
+      ['background-color', 'box-shadow', 'border-color', 'color'],
+      {
+        duration: tokens.sys.motion.duration.short3,
+      },
+    ),
+    fontFamily: tokens.sys.typescale.label.large.family,
+    fontWeight: tokens.sys.typescale.label.large.weight,
+    fontSize: theme.typography.pxToRem(theme.sys.typescale.label.large.size), // the pxToRem should be moved to typescale in the future
+    lineHeight: `calc(${tokens.sys.typescale.label.large.lineHeight} / ${theme.sys.typescale.label.large.size})`,
+    borderRadius: `var(--Button-radius, ${borderRadius})`,
+    backgroundColor: containerColor[ownerState.variant ?? 'text'],
+    color: labelTextColor[ownerState.variant ?? 'text'],
+    boxShadow: containerElevation[ownerState.variant ?? 'text'],
+    // Outlined variant
+    ...(ownerState.variant === 'outlined' && {
+      border: `1px solid ${tokens.sys.color.outline}`,
+      padding: '9px 23px',
+    }),
+    '--Button-gap': '0.5rem',
+    // Sizes are not specified in Material Design 3, this need to be revised
+    ...(ownerState.size === 'small' && {
+      '--Button-gap': '0.45rem',
+      fontSize: theme.typography.pxToRem(theme.sys.typescale.label.large.size - 1), // the pxToRem should be moved to typescale in the future
+      padding: '8px 20px',
+      ...(ownerState.variant === 'outlined' && {
+        padding: '7px 19px',
+      }),
+    }),
+    ...(ownerState.size === 'large' && {
+      '--Button-gap': '0.55rem',
+      fontSize: theme.typography.pxToRem(theme.sys.typescale.label.large.size + 1), // the pxToRem should be moved to typescale in the future
+      padding: '12px 26px',
+      ...(ownerState.variant === 'outlined' && {
+        padding: '11px 25px',
+      }),
+    }),
+    '&:hover': {
+      '--md-comp-button-icon-color': 'var(--md-comp-button-hovered-icon-color)',
+      backgroundColor: hoveredContainerColor[ownerState.variant ?? 'text'],
+      boxShadow: hoveredContainerElevation[ownerState.variant ?? 'text'],
+    },
+    '&:active': {
+      '--md-comp-button-icon-color': 'var(--md-comp-button-pressed-icon-color)',
+      ...((ownerState.disableRipple || ownerState.disableTouchRipple) && {
+        backgroundColor: pressedContainerColor[ownerState.variant ?? 'text'],
+      }),
+      boxShadow: pressedContainerElevation[ownerState.variant ?? 'text'],
+    },
+    [`&.${buttonBaseClasses.focusVisible}`]: {
+      '--md-comp-button-icon-color': 'var(--md-comp-button-focused-icon-color)',
+      backgroundColor: focusedContainerColor[ownerState.variant ?? 'text'],
+      boxShadow: focusedContainerElevation[ownerState.variant ?? 'text'],
+    },
+    [`&.${buttonBaseClasses.disabled}`]: {
+      // Allows developer to specify the disabled icon color var
+      '--md-comp-button-icon-color': 'var(--md-comp-button-disabled-icon-color)',
+      color: disabledLabelTextColor,
+      backgroundColor: disabledContainerColor[ownerState.variant ?? 'text'],
+      boxShadow: tokens.sys.elevation[0],
+      ...(ownerState.variant === 'outlined' && {
+        border: `1px solid ${
+          theme.vars
+            ? `rgba(${theme.vars.sys.color.onSurfaceChannel} / 0.12)`
+            : alpha(theme.sys.color.onSurface, 0.12)
+        }`,
+      }),
+    },
+  };
+});
+
+const ButtonStartIconMD3 = styled('span', {
+  name: 'MuiButton',
+  slot: 'StartIcon',
+  overridesResolver: (props, styles) => {
+    const { ownerState } = props;
+
+    return [styles.startIcon, styles[`iconSize${capitalize(ownerState.size)}`]];
+  },
+})({
+  display: 'inherit',
+  marginRight: 'var(--Button-gap)',
+  ...commonIconStyles,
+});
+
+const ButtonEndIconMD3 = styled('span', {
+  name: 'MuiButton',
+  slot: 'EndIcon',
+  overridesResolver: (props, styles) => {
+    const { ownerState } = props;
+
+    return [styles.endIcon, styles[`iconSize${capitalize(ownerState.size)}`]];
+  },
+})({
+  display: 'inherit',
+  marginLeft: 'var(--Button-gap)',
+  ...commonIconStyles,
+});
+
+const useIsMd3 = () => {
+  const theme = useTheme();
+  return !!theme.sys;
+}
+
 const Button = React.forwardRef(function Button(inProps, ref) {
   // props priority: `inProps` > `contextProps` > `themeDefaultProps`
   const contextProps = React.useContext(ButtonGroupContext);
   const buttonGroupButtonContextPositionClassName = React.useContext(ButtonGroupButtonContext);
   const resolvedProps = resolveProps(contextProps, inProps);
   const props = useDefaultProps({ props: resolvedProps, name: 'MuiButton' });
+  const isMd3 = useIsMd3();
   const {
     children,
     color = 'primary',
@@ -384,22 +698,26 @@ const Button = React.forwardRef(function Button(inProps, ref) {
 
   const classes = useUtilityClasses(ownerState);
 
+  const Root = isMd3 ? ButtonRootMD3 : ButtonRoot;
+  const StartIcon = isMd3 ? ButtonStartIconMD3 : ButtonStartIcon;
+  const EndIcon = isMd3 ? ButtonEndIconMD3 : ButtonEndIcon;
+
   const startIcon = startIconProp && (
-    <ButtonStartIcon className={classes.startIcon} ownerState={ownerState}>
+    <StartIcon className={classes.startIcon} ownerState={ownerState}>
       {startIconProp}
-    </ButtonStartIcon>
+    </StartIcon>
   );
 
   const endIcon = endIconProp && (
-    <ButtonEndIcon className={classes.endIcon} ownerState={ownerState}>
+    <EndIcon className={classes.endIcon} ownerState={ownerState}>
       {endIconProp}
-    </ButtonEndIcon>
+    </EndIcon>
   );
 
   const positionClassName = buttonGroupButtonContextPositionClassName || '';
 
   return (
-    <ButtonRoot
+    <Root
       ownerState={ownerState}
       className={clsx(contextProps.className, classes.root, className, positionClassName)}
       component={component}
@@ -414,7 +732,7 @@ const Button = React.forwardRef(function Button(inProps, ref) {
       {startIcon}
       {children}
       {endIcon}
-    </ButtonRoot>
+    </Root>
   );
 });
