--- packages/mui-material/src/ButtonNew/ButtonNew.types.ts
@@ -1,115 +1,0 @@
-import * as React from 'react';
-import {
-  OverridableStringUnion,
-  OverrideProps,
-  OverridableComponent,
-  OverridableTypeMap,
-} from '@mui/types';
-import { SxProps } from '../styles/Theme.types';
-import { ButtonClasses } from './buttonNewClasses';
-
-export interface ButtonPropsVariantOverrides {}
-
-export interface ButtonPropsColorOverrides {}
-
-export interface ButtonPropsSizeOverrides {}
-
-export interface ButtonActions {
-  focusVisible(): void;
-}
-
-export type ButtonTypeMap<
-  AdditionalProps = {},
-  DefaultComponent extends React.ElementType = 'button',
-> = {
-  props: AdditionalProps & {
-    /**
-     * The content of the component.
-     */
-    children?: React.ReactNode;
-    /**
-     * Override or extend the styles applied to the component.
-     */
-    classes?: Partial<ButtonClasses>;
-    /**
-     * The color of the component.
-     * It supports both default and custom theme colors, which can be added as shown in the
-     * [palette customization guide](https://mui.com/material-ui/customization/palette/#custom-colors).
-     * @default 'primary'
-     */
-    color?: OverridableStringUnion<'primary' | 'secondary' | 'tertiary', ButtonPropsColorOverrides>;
-    /**
-     * If `true`, no elevation is used.
-     * @default false
-     */
-    disableElevation?: boolean;
-    /**
-     * If `true`, the ripple effect is disabled.
-     * @default false
-     */
-    disableRipple?: boolean;
-    /**
-     * If `true`, the touch ripple effect is disabled.
-     * @default false
-     */
-    disableTouchRipple?: boolean;
-    /**
-     * Element placed after the children.
-     */
-    endIcon?: React.ReactNode;
-    /**
-     * If `true`, the button will take up the full width of its container.
-     * @default false
-     */
-    fullWidth?: boolean;
-    /**
-     * The size of the component.
-     * `small` is equivalent to the dense button styling.
-     * @default 'medium'
-     */
-    size?: OverridableStringUnion<'small' | 'medium' | 'large', ButtonPropsSizeOverrides>;
-    /**
-     * Element placed before the children.
-     */
-    startIcon?: React.ReactNode;
-    /**
-     * The system prop that allows defining system overrides as well as additional CSS styles.
-     */
-    sx?: SxProps;
-    /**
-     * @default 0
-     */
-    tabIndex?: NonNullable<React.HTMLAttributes<any>['tabIndex']>;
-    /**
-     * The variant to use.
-     * @default 'text'
-     */
-    variant?: OverridableStringUnion<
-      'text' | 'outlined' | 'filled' | 'filledTonal' | 'elevated',
-      ButtonPropsVariantOverrides
-    >;
-  };
-  defaultComponent: DefaultComponent;
-};
-
-export interface ButtonOwnerState extends ButtonProps {}
-
-/**
- * A utility to create component types that inherit props from the Button.
- * This component has an additional overload if the `href` prop is set which
- * can make extension quite tricky
- */
-export interface ExtendButtonTypeMap<TypeMap extends OverridableTypeMap> {
-  props: TypeMap['props'] & ButtonTypeMap['props'];
-  defaultComponent: TypeMap['defaultComponent'];
-}
-
-export type ExtendButton<TypeMap extends OverridableTypeMap> = ((
-  props: { href: string } & OverrideProps<ExtendButtonTypeMap<TypeMap>, 'a'>,
-) => JSX.Element) &
-  OverridableComponent<ExtendButtonTypeMap<TypeMap>> & { propTypes?: any };
-
-export type ButtonProps<
-  RootComponent extends React.ElementType = ButtonTypeMap['defaultComponent'],
-  AdditionalProps = {},
-> = OverrideProps<ButtonTypeMap<AdditionalProps, RootComponent>, RootComponent>;
