--- packages/mui-material/src/ButtonNew/index.ts
@@ -1,8 +1,0 @@
-'use client';
-export { default } from './ButtonNew';
-export * from './ButtonNew';
-
-export * from './ButtonNew.types';
-
-export { default as buttonClasses } from './buttonNewClasses';
-export * from './buttonNewClasses';
