+++ packages/mui-material/src/themes/md2.js
@@ -1,0 +1,1 @@
+export { default as extendTheme } from '../styles/extendThemeMD2';
