+++ packages/mui-material/src/themes/md3.js
@@ -1,0 +1,1 @@
+export { default as extendTheme } from '../styles/extendTheme';
